ポルノサイトのURLが取得できるようになった

+++ backend/pornblocker-backend/src/main/java/com/pornblocker/pornblockerbackend/domain/model/PornSite.java
@@ -9,7 +9,7 @@ import lombok.NoArgsConstructor;
 @AllArgsConstructor
 @NoArgsConstructor
 public class PornSite {
-  private Integer pornSiteId;
+  private Long pornSiteId;
   private String siteUrl;
 
   public PornSite(PornSiteEntity entity) {
